requirements.txt
cycler==0.11.0
dataclasses
joblib==1.1.0
kiwisolver==1.3.1
lmdb==1.3.0
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.4.0.46
pandas
Pillow==8.4.0
plotly==4.13.0
py-lz4framed==0.14.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
retrying==1.3.3
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
torch==1.7.1
torchsummary==1.5.1
torchvision==0.8.2
tqdm==4.54.0
typing_extensions==4.0.1